requirements.txt
gradio==4.44.1
ipykernel==6.29.5
ipython==8.27.0
jupyter_client==7.4.9
openai==1.93.0
pandas==2.2.2
pillow==10.4.0
torch==2.4.1
transformers==4.44.2
PyYAML==6.0.1
ansi2html==1.9.2
sentence_transformers==5.1.0
nbformat==5.10.4
pydantic==2.9.1
scikit-learn==1.5.1
statsmodels==0.14.4
openpyxl
python-dotenv==1.0.0

# Hugging Face 資料集分析相關套件
datasets>=2.14.0
opencc-python-reimplemented>=0.1.7
numpy>=1.24.0
scipy>=1.10.0